rewrote templates to render functions
